requirements.txt
fastapi
uvicorn
face_recognition
opencv-python
numpy
requests
pillow
mysql-connector-python
python-dotenv